+++ src/models/Leaf.ts
@@ -1,14 +1,14 @@
 import {index, prop} from '@typegoose/typegoose';
-import {HexStringWith0x} from '../types/HexStringWith0x';
+import {HexStringWith0x} from '../types/custom';
 
 @index({feedId: 1, timestamp: 1})
-@index({blockHeight: 1})
+@index({blockId: 1})
 class Leaf {
   @prop()
   _id!: string;
 
   @prop()
-  blockHeight?: number;
+  blockId?: number;
 
   @prop()
   timestamp!: Date;