+++ src/services/BlockMinter.ts
@@ -1,7 +1,7 @@
 import {Logger} from 'winston';
 import {inject, injectable} from 'inversify';
 import {ethers, Signature} from 'ethers';
-import {converters} from '@umb-network/toolbox';
+import {ABI, LeafKeyCoder, LeafValueCoder} from '@umb-network/toolbox';
 import {getModelForClass} from '@typegoose/typegoose';
 
 import ConsensusRunner from './ConsensusRunner';
@@ -17,6 +17,9 @@ import Block from '../models/Block';
 import {ChainStatus} from '../types/ChainStatus';
 import {Consensus} from '../types/Consensus';
 import Settings from '../types/Settings';
+import {LogMint, LogVoter} from '../types/events';
+import {chainReadyForNewBlock} from '../utils/mining';
+import {MintedBlock} from '../types/MintedBlock';
 
 @injectable()
 class BlockMinter {
@@ -33,60 +36,54 @@ class BlockMinter {
   @inject(RevertedBlockResolver) reveredBlockResolver!: RevertedBlockResolver;
 
   async apply(): Promise<void> {
-    const chainStatus = await this.chainContract.resolveStatus();
+    const [chainAddress, chainStatus] = await this.chainContract.resolveStatus();
 
     if (!this.isLeader(chainStatus)) return;
 
     const dataTimestamp = this.timeService.apply(this.settings.dataTimestampOffsetSeconds);
-    const {nextBlockHeight} = chainStatus;
+    const {nextBlockId} = chainStatus;
 
-    if (!(await this.canMint(chainStatus))) {
+    if (!(await this.canMint(chainStatus, dataTimestamp))) {
       return;
     }
 
     this.logger.info(
-      `Proposing new block for block: ${chainStatus.blockNumber}/${nextBlockHeight.toString()} at ${dataTimestamp}...`,
+      `Proposing new block for block: ${chainStatus.blockNumber}/${nextBlockId.toString()} at ${dataTimestamp}...`,
     );
 
     const validators = this.chainContract.resolveValidators(chainStatus);
 
-    const consensus = await this.consensusRunner.apply(
-      dataTimestamp,
-      nextBlockHeight.toNumber(),
-      validators,
-      chainStatus.staked,
-    );
+    const consensus = await this.consensusRunner.apply(dataTimestamp, nextBlockId, validators, chainStatus.staked);
 
     if (!consensus) {
       this.logger.warn(
-        `No consensus for block ${chainStatus.blockNumber}/${chainStatus.nextBlockHeight} at  ${dataTimestamp}`,
+        `No consensus for block ${chainStatus.blockNumber}/${chainStatus.nextBlockId} at  ${dataTimestamp}`,
       );
       return;
     }
 
     this.logger.info(`Minting a block with ${consensus.signatures.length} signatures...`);
 
-    const tx = await this.mint(
+    const mintedBlock = await this.mint(
       consensus.dataTimestamp,
       consensus.root,
-      consensus.numericFcdKeys,
-      consensus.numericFcdValues,
+      consensus.fcdKeys,
+      consensus.fcdValues,
       consensus.signatures,
     );
 
-    if (tx) {
-      this.logger.info(`Minted in TX ${tx}`);
-      await this.saveBlock(consensus);
+    if (mintedBlock) {
+      const {hash, logMint} = mintedBlock;
+      this.logger.info(`New Block ${logMint.blockId} Minted in TX ${hash}`);
+      await this.saveBlock(chainAddress, consensus, mintedBlock);
     }
   }
 
   private isLeader(chainStatus: ChainStatus): boolean {
-    const {blockNumber, nextLeader, nextBlockHeight} = chainStatus;
+    const {blockNumber, nextBlockId, nextLeader} = chainStatus;
 
     this.logger.info(
-      `Next leader for ${blockNumber}/${nextBlockHeight}: ${nextLeader}, ${
-        nextLeader === this.blockchain.wallet.address
-      }`,
+      `Next leader for ${blockNumber}/${nextBlockId}: ${nextLeader}, ${nextLeader === this.blockchain.wallet.address}`,
     );
 
     return nextLeader === this.blockchain.wallet.address;
@@ -102,33 +99,70 @@ class BlockMinter {
     keys: string[],
     values: number[],
     signatures: string[],
-  ): Promise<string | null> {
+  ): Promise<MintedBlock | null> {
     try {
       const components = signatures.map((signature) => BlockMinter.splitSignature(signature));
 
       const tx = await this.chainContract.submit(
         dataTimestamp,
         root,
-        keys.map(converters.strToBytes32),
-        values.map(converters.numberToUint256),
+        keys.map(LeafKeyCoder.encode),
+        values.map((v) => LeafValueCoder.encode(v)),
         components.map((sig) => sig.v),
         components.map((sig) => sig.r),
         components.map((sig) => sig.s),
       );
 
       const receipt = await tx.wait();
-      return receipt.status == 1 ? receipt.transactionHash : null;
+
+      if (receipt.status !== 1) {
+        return null;
+      }
+
+      const logMint = this.getLogMint(receipt.logs);
+      const logsVoters = this.getLogVoters(receipt.logs);
+
+      if (!logMint) {
+        return null;
+      }
+
+      return {hash: receipt.transactionHash, anchor: receipt.blockNumber, logMint, logsVoters};
     } catch (e) {
       this.logger.error(e);
       return null;
     }
   }
 
-  private async canMint(chainStatus: ChainStatus): Promise<boolean> {
-    if (chainStatus.lastBlockHeight.gt(0) && chainStatus.lastBlockHeight.eq(chainStatus.nextBlockHeight)) {
-      this.logger.info(
-        `skipping ${chainStatus.blockNumber}/${chainStatus.nextBlockHeight.toString()}, waiting for new round`,
-      );
+  private getLogMint(logs: ethers.providers.Log[]): LogMint {
+    const iface = new ethers.utils.Interface(ABI.chainAbi);
+    const logMint = logs.map((log) => iface.parseLog(log)).find((event) => event.name === 'LogMint');
+
+    if (!logMint) {
+      throw Error('can`t find LogMint in logs');
+    }
+
+    const {minter, staked, blockId, power} = logMint.args;
+    return {minter, staked, blockId, power};
+  }
+
+  private getLogVoters(logs: ethers.providers.Log[]): LogVoter[] {
+    const iface = new ethers.utils.Interface(ABI.chainAbi);
+    const logVoters = logs.map((log) => iface.parseLog(log)).filter((event) => event.name === 'LogVoter');
+
+    if (!logVoters) {
+      throw Error('can`t find logVoters in logs');
+    }
+
+    return logVoters.map((logVoter) => {
+      return {vote: logVoter.args.vote, voter: logVoter.args.voter, blockId: logVoter.args.blockId};
+    });
+  }
+
+  private async canMint(chainStatus: ChainStatus, dataTimestamp: number): Promise<boolean> {
+    const [ready, error] = chainReadyForNewBlock(chainStatus, dataTimestamp);
+
+    if (!ready) {
+      this.logger.info(error);
       return false;
     }
 
@@ -138,13 +172,13 @@ class BlockMinter {
       return true;
     }
 
-    await this.reveredBlockResolver.apply(lastSubmittedBlock?.height, chainStatus.nextBlockHeight.toNumber());
+    await this.reveredBlockResolver.apply(lastSubmittedBlock?.blockId, chainStatus.nextBlockId);
     lastSubmittedBlock = await BlockMinter.getLastSubmittedBlock();
 
-    const allowed = lastSubmittedBlock ? chainStatus.nextBlockHeight.gt(lastSubmittedBlock.height) : true;
+    const allowed = lastSubmittedBlock ? chainStatus.nextBlockId > lastSubmittedBlock.blockId : true;
 
+    // TODO this can be removed if we add feature to Chain to start blockId from 1, not from 0.
     if (!allowed) {
-      // this can happen when workers will be scheduled faster than they can mint
       this.logger.info(`you already mined this block`);
     }
 
@@ -152,19 +186,32 @@ class BlockMinter {
   }
 
   private static async getLastSubmittedBlock(): Promise<Block | undefined> {
-    const blocks: Block[] = await getModelForClass(Block).find({}).limit(1).sort({height: -1}).exec();
+    const blocks: Block[] = await getModelForClass(Block).find({}).limit(1).sort({blockId: -1}).exec();
     return blocks[0];
   }
 
-  private async saveBlock(consensus: Consensus): Promise<void> {
+  private async saveBlock(chainAddress: string, consensus: Consensus, mintedBlock: MintedBlock): Promise<void> {
+    const votes: Record<string, string> = {};
+
+    mintedBlock.logsVoters.forEach((logVoter) => {
+      votes[logVoter.voter] = logVoter.vote.toString();
+    });
+
     await this.saveMintedBlock.apply({
+      id: `block::${mintedBlock.logMint.blockId}`,
+      chainAddress,
       dataTimestamp: new Date(consensus.dataTimestamp * 1000),
       timestamp: new Date(),
       leaves: consensus.leaves,
-      blockHeight: consensus.blockHeight,
+      blockId: mintedBlock.logMint.blockId.toNumber(),
+      anchor: mintedBlock.anchor,
       root: consensus.root,
-      numericFcdKeys: consensus.numericFcdKeys,
-      numericFcdValues: consensus.numericFcdValues,
+      fcdKeys: consensus.fcdKeys,
+      fcdValues: consensus.fcdValues,
+      votes: votes,
+      power: mintedBlock.logMint.power.toString(),
+      staked: mintedBlock.logMint.staked.toString(),
+      miner: mintedBlock.logMint.minter,
     });
   }
 }