+++ src/services/SaveMintedBlock.ts
@@ -1,19 +1,25 @@
 import {getModelForClass} from '@typegoose/typegoose';
 import {injectable} from 'inversify';
-import {HexStringWith0x} from '../types/HexStringWith0x';
+import {HexStringWith0x} from '../types/custom';
 import {v4 as uuid} from 'uuid';
 import Block from '../models/Block';
 import Leaf from '../models/Leaf';
 
 type Params = {
   id?: string;
+  chainAddress: string;
   leaves: Leaf[];
-  blockHeight: number;
+  blockId: number;
+  anchor: number;
   root: string;
   dataTimestamp: Date;
   timestamp: Date;
-  numericFcdKeys: string[];
-  numericFcdValues: number[];
+  fcdKeys: string[];
+  fcdValues: number[];
+  votes: Record<string, string>;
+  power: string;
+  staked: string;
+  miner: string;
 };
 
 @injectable()
@@ -21,18 +27,24 @@ class SaveMintedBlock {
   async apply(params: Params): Promise<Block> {
     const block = new Block();
     block._id = params.id || uuid();
+    block.chainAddress = params.chainAddress;
     block.dataTimestamp = params.dataTimestamp;
     block.timestamp = params.timestamp;
-    block.height = params.blockHeight;
+    block.blockId = params.blockId;
+    block.anchor = params.anchor;
     block.root = params.root;
     block.data = this.treeDataFor(params.leaves);
-    block.numericFcdKeys = params.numericFcdKeys;
-    block.numericFcdValues = params.numericFcdValues;
-    await this.attachLeavesToBlockHeight(params.leaves, params.blockHeight);
+    block.fcdKeys = params.fcdKeys;
+    block.fcdValues = params.fcdValues;
+    block.votes = params.votes;
+    block.power = params.power;
+    block.staked = params.staked;
+    block.minter = params.miner;
+    await this.attachLeavesToBlock(params.leaves, params.blockId);
     return await getModelForClass(Block).create(block);
   }
 
-  private async attachLeavesToBlockHeight(leaves: Leaf[], blockHeight: number): Promise<void> {
+  private async attachLeavesToBlock(leaves: Leaf[], blockId: number): Promise<void> {
     return getModelForClass(Leaf)
       .updateMany(
         {
@@ -42,7 +54,7 @@ class SaveMintedBlock {
         },
         {
           $set: {
-            blockHeight: blockHeight,
+            blockId: blockId,
           },
         },
       )