+++ src/types/MintedBlock.ts
@@ -1,0 +1,8 @@
+import {LogMint, LogVoter} from './events';
+
+export type MintedBlock = {
+  hash: string;
+  anchor: number;
+  logMint: LogMint;
+  logsVoters: LogVoter[];
+};