+++ test/services/SaveMintedBlock.test.ts
@@ -32,19 +32,25 @@ describe('SaveMintedBlock', () => {
 
   it("builds tree data and saves in block's object", async () => {
     const leaves: Leaf[] = [
-      {_id: uuid.v4(), label: 'ETH-USD', valueBytes: '0x01', timestamp: new Date(), blockHeight: 1},
-      {_id: uuid.v4(), label: 'USD-ETH', valueBytes: '0x02', timestamp: new Date(), blockHeight: 1},
+      {_id: uuid.v4(), label: 'ETH-USD', valueBytes: '0x01', timestamp: new Date(), blockId: 1},
+      {_id: uuid.v4(), label: 'USD-ETH', valueBytes: '0x02', timestamp: new Date(), blockId: 1},
     ];
 
     const result = await saveMintedBlock.apply({
       id: 'block::1',
-      blockHeight: 1,
+      chainAddress: '0x333',
+      blockId: 1,
       leaves,
-      numericFcdKeys: ['ETH-USD', 'USD-ETH'],
-      numericFcdValues: [1337.12, 512.12],
+      fcdKeys: ['ETH-USD', 'USD-ETH'],
+      fcdValues: [1337.12, 512.12],
       root: '0x00',
       timestamp: new Date(1000),
       dataTimestamp: new Date(1000),
+      power: '2',
+      staked: '1',
+      votes: {'0xab': '1'},
+      anchor: 123,
+      miner: '0x111',
     });
 
     expect(result.data).to.be.an('object').that.deep.eq({
@@ -55,24 +61,34 @@ describe('SaveMintedBlock', () => {
 
   it('generates UUID and saves the block to database', async () => {
     await saveMintedBlock.apply({
-      blockHeight: 1,
+      chainAddress: '0x333',
+      blockId: 1,
       leaves: [],
-      numericFcdKeys: ['ETH-USD', 'USD-ETH'],
-      numericFcdValues: [1337.12, 512.12],
+      fcdKeys: ['ETH-USD', 'USD-ETH'],
+      fcdValues: [1337.12, 512.12],
       root: '0x00',
       dataTimestamp: new Date(1000),
       timestamp: new Date(1000),
+      power: '2',
+      staked: '1',
+      votes: {'0xab': '1'},
+      anchor: 123,
+      miner: '0x1111',
     });
 
     const blockFromDb = await getModelForClass(Block).findOne();
     expect(blockFromDb).to.be.an('object');
     expect(uuid.validate(blockFromDb?._id as string)).to.be.true;
-    expect(blockFromDb?.height).to.be.eq(1);
-    expect(blockFromDb?.numericFcdKeys).to.be.deep.eq(['ETH-USD', 'USD-ETH']);
-    expect(blockFromDb?.numericFcdValues).to.be.deep.eq([1337.12, 512.12]);
+    expect(blockFromDb?.blockId).to.be.eq(1);
+    expect(blockFromDb?.fcdKeys).to.be.deep.eq(['ETH-USD', 'USD-ETH']);
+    expect(blockFromDb?.fcdValues).to.be.deep.eq([1337.12, 512.12]);
     // is undefined or an empty object
     expect(blockFromDb?.data).to.satisfies((data: any) => data === undefined || Object.keys(data).length === 0);
     expect(blockFromDb?.root).to.be.eq('0x00');
     expect(blockFromDb?.timestamp).to.be.a('Date');
+    expect(blockFromDb?.anchor).to.eq(123);
+    expect(blockFromDb?.power).to.eq('2');
+    expect(blockFromDb?.staked).to.eq('1');
+    expect(blockFromDb?.votes).to.eql({'0xab': '1'});
   });
 });